requirements.txt
python-dotenv==0.21.0
python-telegram==0.16.0
